Renamed header folder to navbar
