+++ src/components/header.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Navbar from "./header/navbar";
+import Navbar from "./navbar/navbar";
 
 const Header = () => {
   return <Navbar />;
